requirements.txt
Flask
transformers
torch
Pillow
requests
pandas
gunicorn
flask-cors
torchao